DPI-1: add semantic round-trip tests for SCP/EICP DPI field
Add send->recv loopback tests for the create-session DPI field in
tests/libipm/test_libipm_recv_calls.c: a valid in-range DPI survives the
round trip and out-of-range/absent values are rejected to 0 on receive
(defense in depth), with the non-DPI fields verified intact. These are
semantic (value-based) round trips, not raw-byte assertions, so they stay
robust against future field additions - matching the repo's convention of
not byte-asserting individual application messages.

Also narrow the local dpi to unsigned short in xrdp_mm.c to match the SCP
API and the rest of the chain, and reword the BACKLOG test note.

make check: libipm 39 (was 35), all suites green. astyle clean.

+++ xrdp/xrdp_mm.c
@@ -313,7 +313,7 @@ xrdp_mm_create_session(struct xrdp_mm *self)
     int rv = 0;
     int xserverbpp;
     enum scp_session_type type;
-    unsigned int dpi = 0;
+    unsigned short dpi = 0;
 
     /* Map the session code to an SCP session type */
     switch (self->code)
